% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{get_clean_field_plan}
\alias{get_clean_field_plan}
\title{Read and clean the field plan data}
\usage{
get_clean_field_plan(field_plan_sheet)
}
\arguments{
\item{field_plan_sheet}{A data frame containing the raw field plan data}
}
\value{
A cleaned data frame with deployment and retrieval statistics
}
\description{
This function processes the field plan data by selecting relevant columns, renaming them, calculating deployment and retrieval success rates,
and merging with historical data to compute previous deployments. A connection to the SEATRACK database is required to calculate retrieval success.
}
\examples{
\dontrun{
field_plan_sheet <- get_field_plan("path/to/fieldplan.xlsx")
seatrackR::connectSeatrack()
field_plan_clean <- get_clean_field_plan(field_plan_sheet)
}
}
\concept{field_planning}
