% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_location_unprocessed}
\alias{get_location_unprocessed}
\title{Get unimported metadata}
\usage{
get_location_unprocessed(location)
}
\arguments{
\item{location}{Character string specifying the location (colony) to check for unprocessed metadata.}
}
\description{
Checks a location to see if there is unimported metadata in the "not_processed" folder.
If found, returns a list of file paths to these unimported metadata files.
}
\concept{metadata}
