% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_logger_from_metadata}
\alias{get_logger_from_metadata}
\title{Find logger instances in files}
\usage{
get_logger_from_metadata(logger_id, all_master_import_list = NULL)
}
\arguments{
\item{logger_id}{logger ID of desired logger}

\item{all_master_import_list}{list of master import sheets, as returned by \code{load_all_master_import(combine = FALSE)}. If not provided, this will be generated inside the function}
}
\value{
list of all matches, where each element of the list is a list containing the row of data, the file path and the index.
}
\description{
This function tries to find a logger ID in all master import files. You can either provide an already loaded list of master import sheets, or the function will load them itself.
}
