% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_prev_session.R
\name{get_raw_data_prev_session}
\alias{get_raw_data_prev_session}
\title{Get raw data from previous sessions}
\usage{
get_raw_data_prev_session(
  logger_serial_nos = c(),
  zip_filename = paste0("raw_data_", Sys.Date(), ".zip"),
  search_path = NULL
)
}
\arguments{
\item{logger_serial_nos}{A character vector of logger serial numbers to get previous session data for.}

\item{zip_filename}{The name of the output zip file. Defaults to "raw_data_<current_date>.zip".}

\item{search_path}{The path to search for raw data files. If NULL, defaults to the "ALL" folder in the SeaTrack database imports directory.}
}
\value{
None, but exports a zip file containing the raw data files and a metadata CSV.
}
\description{
This function gets any sessions occuring before the latest session of a seatrack GLS logger and
collects the associated raw data files into a zip file. The zip file also contains a CSV
with metadata about the previous sessions.
}
\details{
This history is useful when a manufacturer is attempting to recover data from a nonresponsive logger.
}
\concept{nonresponsive}
