% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_metadata}
\alias{gls_metadata}
\title{Get metadata from database based on GLS files in import directory}
\usage{
gls_metadata(import_directory, no_pos_only = TRUE)
}
\arguments{
\item{import_directory}{Path to the directory containing GLS files.}

\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}
}
\value{
A dataframe containing metadata for the GLS loggers found in the import directory.
}
\description{
Function to scan a directory for GLS files and retrieve corresponding metadata from the Sea Track database.
}
\concept{gls_helper}
