% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_prepare_calibration}
\alias{gls_prepare_calibration}
\title{Prepare GLS calibration data using seatrack database}
\usage{
gls_prepare_calibration(import_directory, output_directory, no_pos_only = TRUE)
}
\arguments{
\item{import_directory}{Path to the directory containing GLS files.}

\item{output_directory}{Path to the directory where the prepared calibration data will be saved.}

\item{no_pos_only}{Logical indicating whether to include only loggers without position data in the database. Default is TRUE.}
}
\value{
None. The function saves the prepared calibration data to the specified output directory.
}
\description{
Function to prepare GLS calibration data for use with seatrackRgls, based on GLS files in the import directory and metadata from the Sea Track database.
}
\concept{gls_helper}
