% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_process_positions}
\alias{gls_process_positions}
\title{Process GLS position data using seatrackRgls}
\usage{
gls_process_positions(import_directory, calibration_data, output_directory)
}
\arguments{
\item{import_directory}{Path to the directory containing GLS files.}

\item{calibration_data}{Dataframe containing GLS calibration data.}

\item{output_directory}{Path to the directory where the processed position data will be saved.}
}
\value{
None. The function saves the processed position data to the specified output directory.
}
\description{
Function to process GLS position data from a specified directory using seatrackRgls.
The only additional functionality is fetching colony information from the Sea Track database.
}
\concept{gls_helper}
