% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_seatrack_calibration}
\alias{gls_seatrack_calibration}
\title{Handle existing GLS calibration data}
\usage{
gls_seatrack_calibration(metadata_path)
}
\arguments{
\item{metadata_path}{Path to metadata Excel file}
}
\value{
A list with two dataframes: calibration_data and extra_metadata
}
\description{
Function to reshape existing seatrack GLS calibration settings for immediate use in seatrackRgls
}
\concept{gls_helper}
