% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gls_functions.R
\name{gls_seatrack_colony_info}
\alias{gls_seatrack_colony_info}
\title{Get \code{all_colony_info} for GLS processing}
\usage{
gls_seatrack_colony_info()
}
\description{
Simple function to get seatrack colonies from database and shape the dataframe for immediate use in seatrackRgls
}
\concept{gls_helper}
