% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_all_master_import}
\alias{load_all_master_import}
\title{Load master import file for all colonies}
\usage{
load_all_master_import(combine = TRUE)
}
\arguments{
\item{combine}{Boolean determining whether or not to combine the sheets into a single dataframe.}
}
\value{
If combine is TRUE: A list consisting of combined metadata and startup_shutdown sheets, with an extra column for path appended to each.
Otherwise a list where every element is a sheet
}
\description{
This function attempts to load all master import sheets available in the seatrack folder.
}
\concept{metadata}
