% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_master_import}
\alias{load_master_import}
\title{Load master import file for a given colony}
\usage{
load_master_import(colony = NULL, file_path = NULL, use_stored = TRUE)
}
\arguments{
\item{colony}{A character string specifying the name of the colony.}

\item{file_path}{File path of master import file}

\item{use_stored}{If TRUE, use a pre-existing path, rather than searching for a new one. Defaults to TRUE.}
}
\value{
A LoadedWB object.
}
\description{
This function loads the master import file for a specified colony or directly from a file path.
Either a colony name of a file path must be provided.
It iterates through the appropriate sheets and combines the data into a list of data frames.
}
\examples{
\dontrun{
load_master_import("ColonyA")
}
}
\concept{metadata}
