% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{save_all_modified}
\alias{save_all_modified}
\title{Save all modified master sheets}
\usage{
save_all_modified(all_master_sheets)
}
\arguments{
\item{all_master_sheets}{LoadedWBCollection object.}
}
\description{
This function will save any modified master sheets from a LoadedWBCollection object.
}
\concept{metadata}
